requirements.txt
requests==2.32.3
pytest==8.4.0
jsonschema==4.24.0
faker==37.4.0
psycopg2-binary==2.9.10